refactor the PropertyEditor to be more extensive
